refactor: move client-side optimizations into dedicated ClientShell component
